Add post-deployment rules for PXE network configuration
Fuelmenu should allow post-deployment configuration of
PXE network, but place the following restrictions:
* new_range must contain old_range in dhcp pool
* PXE interface must remain the same

For all intents and purposes, the changes are limited to
extending the range from the previously configured one.
A common use case of this would be if a user defined a /24
network for PXE, and then decided to change it to a /23
network. This would meet the user's needs and not impact
existing deployed nodes.

Change-Id: I42fc5d3261fe7e6c4a996e9895710970fc721247
Closes-Bug: #1503828