Disable dhcp_nowait and fail on DHCP if none detected
dhcp_nowait was removed a long time ago from L23network
puppet module, so it cannot be used as an option for enabling
DHCP when there are no available DHCP servers. Since we are
checking for the presence of DHCP servers, we should fail to
apply configuration instead of waiting for Puppet to fail.

Change-Id: Ia8275d32b13a22d56922cd4a6fd6528bb7c77692
Closes-Bug: #1510059