Disable --save-only mode during post-deployment
It is common for users to run bootstrap_admin_node to apply
a patch to Fuel Master, which has the side effect of calling
fuelmenu --save-only --iface=ethN. This behavior initializes
all fields to defaults, which has negative consequences.

Now fuelmenu --save-only --iface=ethN will result in a message
saying that the user should run full fuelmenu in order to make
changes.

Change-Id: I78d3261a5e2700b7533d9873f6bec9cc4e98c362
Closes-Bug: #1513486