Unhardcode default interface from eth0
Now the first detected physical interface will be
used as the default network interface. If for any
reason no physical interfaces are detected, fuelmenu
will exit with an error.

New common.network method getPhysicalIfaces() returns
a sorted list of interfaces that are physical

New common.network method isPhyiscal() returns true
if the string 'virtual' is not in its sysfs realpath.

Change-Id: I8d5f0fbcc1426a94b433d1a097fb51cc6915c154
Closes-Bug: #1512479