Do not update console loglevel
The flow of kernel messages breaks the look of fuel-menu. To work around
it, fuel-menu updates the kernel.printk sysctl to suppress them being
printed to the console (restoring the original value upon the exit).

A better way is updating this variable outside fuel-menu so more
components could benefit from it.

This patch removes the code that changes printk from fuel-menu.

Change-Id: I62dd9c1db19cc9f3a74fd08c80d441ed8cf22692
Depends-On: If6c65a08c83e9ff5b4a828e2a45c34c78e571e72
Closes-Bug: #1535419