Ensure network interface is up befure running DHCP server discovery.
It happened that the interface where DHCP servers are about to discover
was always up due to NetworkManager service. Since we are disabling it
we need to ensure that the interface is up.
Also there is way more neat way to delete default routes which doesn't
break when we have more than one of them.

Change-Id: Ieda8b67be8133bdc953ace2054aaa6db1ce949c0
Closes-Bug: #1543767