Use first NIC with IP address by default
Previously fuelmenu took first network interface
(if it's not specified by user) as default option
for admin NIC whether it has IP address or not.
Now it tries to find a NIC with IP assigned, so
it works fine without additional options when
admin network is attached to not first interface.

Change-Id: Iabd0ebb81af40b4f97f67df92403fcff7c268edf
Related-bug: #1566557